pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tunetables"
version = "0.2"
authors = [
{name="Benjamin Feuer"},
{name="Niv Cohen"},
{name="Colin White"},
{name="Valeriia Cherepanova"},
]
description = "TuneTables is a tabular classification framework that extends the capabilities of pretrained prior-fitted networks like TabPFN'"
readme = "README.md"
requires-python = ">=3.7"
dependencies=[
        'torch>=1.9.0',
        'scikit-learn>=0.24.2,<=1.4.0',
        'pyyaml>=5.4.1,<=6.0.1',
        'numpy>=1.21.2,<=1.26.4',
        'requests>=2.23.0',
        'uncertainty_metrics',
        'wandb==0.16.1',
        'tqdm>=4.62.1',
        'configspace==0.4.21',
        'pandas>=1.3.3,<=2.2.0',
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.hatch.build]
exclude = [
  "*.cpkt",
  "tunetables/data/*",
  "tunetables/logs/*",
  "tunetables/config/*",
  "tunetables/wandb/*",
  "tunetables/datasets/*.pickle",
]

[project.optional-dependencies]
full = [
    'gpytorch>=1.5.0', # training
    'catboost>=0.26.1', # baselines
    'auto-sklearn>=0.14.5', # baselines
    'xgboost>=1.4.0', # baselines
    'hyperopt>=0.2.5', # baselines
    'configspace>=0.4.21', # baselins + training + evaluation
    'openml>=0.12.2', # evaluation + baselines
    'seaborn==0.11', # evaluation
]

[project.urls]
"Homepage" = "https://github.com/penfever/TuneTables/"
